This wil add the middleware to the routes

+++ routes/web.php
@@ -75,11 +75,11 @@ Route::post('/update-password', [UserController::class, 'updatePassword'])->name
 
 
 Route::get('/admin/products/create', [ProductController::class, 'create'])->name('admin.products.create')->middleware('super_admin')->middleware('auth');
-Route::post('/admin/products', [ProductController::class, 'store'])->name('admin.products.store');
-Route::get('/admin/products/{id}/edit', [ProductController::class, 'edit'])->name('admin.products.edit');
-Route::put('/admin/products/{id}', [ProductController::class, 'update'])->name('admin.products.update');
-Route::delete('/admin/products/{id}', [ProductController::class, 'destroy'])->name('admin.products.destroy');
-Route::get('/admin/dashboard', [ProductController::class, 'dashboard'])->name('dashboard');
+Route::post('/admin/products', [ProductController::class, 'store'])->name('admin.products.store')->middleware('super_admin')->middleware('auth');
+Route::get('/admin/products/{id}/edit', [ProductController::class, 'edit'])->name('admin.products.edit')->middleware('super_admin')->middleware('auth');
+Route::put('/admin/products/{id}', [ProductController::class, 'update'])->name('admin.products.update')->middleware('super_admin')->middleware('auth');
+Route::delete('/admin/products/{id}', [ProductController::class, 'destroy'])->name('admin.products.destroy')->middleware('super_admin')->middleware('auth');
+Route::get('/admin/dashboard', [ProductController::class, 'dashboard'])->name('dashboard')->middleware('super_admin')->middleware('auth');
 Route::get('/get-product-description', [ProductController::class, 'getProductDescription'])->name('get-product-description');
 Route::get('/get-product-quantity', [ProductController::class, 'getStockQuantity'])->name('get-product-quantity');
 Route::get('/admin/allproducts', [ProductController::class, 'allproducts'])->name('admin.viewproducts')->middleware('super_admin')->middleware('auth');
